Refactoring of components and adding in conditional views
